Do not save files by title
The title on SoundCloud for the song "Sorority" by Lucy Bedroque is "𝟎𝟐. SORORITY .*♱ ##𝔰𝔦𝔰𝔱𝔢𝔯𝔥𝔬𝔬𝔡". As a result, when previously downloading the audio file, the filename would be "𝟎𝟐. SORORITY .*♱ ##𝔰𝔦𝔰𝔱𝔢𝔯𝔥𝔬𝔬𝔡.m4a", which would break playback. This changes it to use the uploader id and id in the filename instead, which won't cause these weird filenames.

For the sake of being explicit: you may think the issue is because the filename had spaces, but this is *not* the case: filenames with spaces are fine.